restore test to minimize spurious changes
