+++ packages/jsii-calc/lib/index.ts
@@ -3,5 +3,6 @@ export * from './compliance';
 export * from './documented';
 export * from './erasures';
 export * from './stability';
+export * from './submodules';
 
 export * as submodule from './submodule';
